+++ checks/with-build-script-custom/src/main.rs
@@ -1,0 +1,4 @@
+const DATA: &str = include_str!("../target/mydata");
+fn main() {
+    println!("{DATA}");
+}